Added support to build implants for unsupported platforms
