+++ implant/sliver/handlers/generic-tun-handlers.go
@@ -1,3 +1,5 @@
+// +build windows linux darwin
+
 package handlers
 
 /*
