+++ implant/sliver/limits/limits_default.go
@@ -1,0 +1,30 @@
+// +build !windows !linux !darwin
+
+package limits
+
+/*
+	Sliver Implant Framework
+	Copyright (C) 2019  Bishop Fox
+
+	This program is free software: you can redistribute it and/or modify
+	it under the terms of the GNU General Public License as published by
+	the Free Software Foundation, either version 3 of the License, or
+	(at your option) any later version.
+
+	This program is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+	GNU General Public License for more details.
+
+	You should have received a copy of the GNU General Public License
+	along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
+func isDomainJoined() (bool, error) {
+	return false, nil
+}
+
+// PlatformLimits - No platform limits for default
+func PlatformLimits() {
+
+}
