+++ implant/sliver/version/version_default.go
@@ -1,0 +1,6 @@
+package version
+
+// GetVersion returns the os version information
+func GetVersion() string {
+	return ""
+}
