+++ server/gogo/go.go
@@ -39,12 +39,43 @@ var (
 
 	// ValidCompilerTargets - Supported compiler targets
 	ValidCompilerTargets = map[string]bool{
-		"darwin/386":    true,
-		"darwin/amd64":  true,
-		"linux/386":     true,
-		"linux/amd64":   true,
-		"windows/386":   true,
-		"windows/amd64": true,
+		"aix/ppc64":       true,
+		"darwin/arm64":    true,
+		"darwin/amd64":    true,
+		"dragonfly/amd64": true,
+		"freebsd/386":     true,
+		"freebsd/amd64":   true,
+		"freebsd/arm":     true,
+		"freebsd/arm64":   true,
+		"illumos/amd64":   true,
+		// "js/wasm":         true, // This is probably the most problimatic
+		"linux/386":      true,
+		"linux/amd64":    true,
+		"linux/arm":      true,
+		"linux/arm64":    true,
+		"linux/ppc64":    true,
+		"linux/ppc64le":  true,
+		"linux/mips":     true,
+		"linux/mipsle":   true,
+		"linux/mips64":   true,
+		"linux/mips64le": true,
+		"linux/riscv64":  true,
+		"linux/s390x":    true,
+		"netbsd/386":     true,
+		"netbsd/amd64":   true,
+		"netbsd/arm":     true,
+		"netbsd/arm64":   true,
+		"openbsd/386":    true,
+		"openbsd/amd64":  true,
+		"openbsd/arm":    true,
+		"openbsd/arm64":  true,
+		"plan9/386":      true,
+		"plan9/amd64":    true,
+		"plan9/arm":      true,
+		"solaris/amd64":  true,
+		"windows/386":    true,
+		"windows/amd64":  true,
+		"windows/arm":    true,
 	}
 )
